+++ packages/ppx_compose.0.0.4/CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## v0.0.4 - 2017-12-17
+
+- Add missing location information in generated code.
+
 ## v0.0.3 - 2017-06-04
 
 - Internal improvements.